refactor(server): split websocket iq handlers
