+++ server/crates/waddle-server/src/server/routes/websocket/handlers/iq/mod.rs
@@ -70,6 +70,26 @@ use waddle_xmpp::{
 };
 use xmpp_parsers::minidom::Element;
 
+mod blocking;
+mod commands;
+mod muc_admin;
+mod push;
+mod roster;
+mod search;
+mod session_misc;
+mod vcard_private;
+
+use blocking::handle_blocking_iq;
+use commands::handle_command_iq;
+use muc_admin::handle_muc_admin_iq;
+use push::handle_push_iq;
+use roster::handle_roster_iq;
+use search::{handle_channel_search_iq, handle_user_search_iq};
+use session_misc::{handle_isr_token_request_iq, handle_muc_self_ping_iq, route_full_jid_iq};
+#[cfg(test)]
+use vcard_private::avatar_vcard_from_user_profile;
+use vcard_private::{handle_private_storage_iq, handle_vcard_iq};
+
 #[cfg(test)]
 use waddle_xmpp::protocol::frame::{parse_frame, InboundFrame};
 
@@ -2364,489 +2384,6 @@ pub async fn handle_iq_with_conn_state(
     )]
 }
 
-async fn handle_muc_self_ping_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    sender_jid: Option<&FullJid>,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-) -> Vec<String> {
-    let Some(sender_jid) = sender_jid else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "auth",
-            "not-authorized",
-        )];
-    };
-    let Some(target) = iq
-        .to
-        .as_ref()
-        .and_then(|jid| jid.clone().try_into_full().ok())
-    else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "modify",
-            "bad-request",
-        )];
-    };
-    let room_jid = target.to_bare();
-    let nick = target.resource().to_string();
-    let Some(room_actor) = get_room_actor(state, &room_jid).await else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "cancel",
-            "item-not-found",
-        )];
-    };
-    match room_actor
-        .ask(PingSelfCheck {
-            nick,
-            sender_jid: sender_jid.clone(),
-        })
-        .await
-    {
-        Ok(()) => vec![build_iq_result_xml(
-            &iq.id,
-            response_from,
-            response_to,
-            None,
-        )],
-        Err(kameo::error::SendError::HandlerError(_)) => vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "cancel",
-            "not-acceptable",
-        )],
-        Err(error) => {
-            warn!(room = %room_jid, error = ?error, "Failed to process MUC self-ping");
-            vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "wait",
-                "internal-server-error",
-            )]
-        }
-    }
-}
-
-async fn handle_isr_token_request_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    authenticated_session: &Option<Session>,
-    sender_jid: Option<&FullJid>,
-) -> Vec<String> {
-    let (Some(session), Some(sender_jid)) = (authenticated_session.as_ref(), sender_jid) else {
-        return vec![iq_to_xml(build_isr_token_error(iq, "not-authorized"))];
-    };
-    let token: IsrToken = state
-        .deps
-        .protocol
-        .isr_token_store
-        .create_token(session.user_id.to_string(), sender_jid.to_bare());
-    vec![iq_to_xml(build_isr_token_result(iq, &token))]
-}
-
-async fn route_full_jid_iq(
-    mut iq: xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    sender_jid: Option<&FullJid>,
-    target: FullJid,
-    response_from: Option<&str>,
-) -> Vec<String> {
-    let Some(sender_jid) = sender_jid else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            None,
-            "auth",
-            "not-authorized",
-        )];
-    };
-    let blocking = DatabaseBlockingStorage::new(state.deps.app_state.db_pool.global().clone());
-    match blocking
-        .is_blocked(&target.to_bare(), &sender_jid.to_bare())
-        .await
-    {
-        Ok(true) => {
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                Some(sender_jid.as_str()),
-                "cancel",
-                "service-unavailable",
-            )];
-        }
-        Ok(false) => {}
-        Err(error) => {
-            warn!(error = %error, target = %target, sender = %sender_jid, "Failed to check blocklist before routing IQ");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                Some(sender_jid.as_str()),
-                "wait",
-                "internal-server-error",
-            )];
-        }
-    }
-    iq.from = Some(Jid::from(sender_jid.clone()));
-    iq.to = Some(Jid::from(target.clone()));
-    match state
-        .deps
-        .protocol
-        .connection_registry
-        .send_to(&target, Stanza::Iq(iq.clone()))
-        .await
-    {
-        waddle_xmpp::registry::SendResult::Sent => Vec::new(),
-        waddle_xmpp::registry::SendResult::NotConnected
-        | waddle_xmpp::registry::SendResult::ChannelClosed => {
-            let sender = sender_jid.to_string();
-            vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                Some(sender.as_str()),
-                "cancel",
-                "service-unavailable",
-            )]
-        }
-    }
-}
-
-async fn handle_channel_search_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    muc_domain: &str,
-    state: &WebSocketState,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-) -> Vec<String> {
-    if iq
-        .to
-        .as_ref()
-        .is_some_and(|to| to.to_string() != muc_domain)
-    {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "cancel",
-            "item-not-found",
-        )];
-    }
-    let Some(request) = parse_search_request(iq) else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "modify",
-            "bad-request",
-        )];
-    };
-    let limit = request.max.unwrap_or(50).clamp(1, 200) as usize;
-    let channels =
-        match list_xmpp_channels(state.deps.app_state.db_pool.global_actor().clone(), 500, 0).await
-        {
-            Ok(channels) => channels,
-            Err(error) => {
-                warn!(error = %error, "Failed to load channels for WebSocket search");
-                return vec![build_iq_error_xml_with_addresses(
-                    &iq.id,
-                    response_from,
-                    response_to,
-                    "wait",
-                    "internal-server-error",
-                )];
-            }
-        };
-    let mut results = Vec::new();
-    for channel in channels {
-        let Some(room_jid) = waddle_xmpp::managed_room_jid(&channel.id, muc_domain).ok() else {
-            continue;
-        };
-        let result = ChannelResult::new(room_jid.to_string())
-            .with_name(channel.name)
-            .with_description(channel.description.unwrap_or_default());
-        if result.matches_query(&request.query) {
-            results.push(result);
-            if results.len() >= limit {
-                break;
-            }
-        }
-    }
-    vec![iq_to_xml(build_search_response(iq, &results))]
-}
-
-async fn handle_user_search_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    domain: &str,
-    state: &WebSocketState,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-) -> Vec<String> {
-    match &iq.payload {
-        xmpp_parsers::iq::IqType::Get(_) => {
-            let payload = Element::builder("query", "jabber:iq:search")
-                .append(
-                    Element::builder("instructions", "jabber:iq:search")
-                        .append("Search users by username or email")
-                        .build(),
-                )
-                .append(Element::builder("nick", "jabber:iq:search").build())
-                .append(Element::builder("email", "jabber:iq:search").build())
-                .build();
-            vec![build_iq_result_xml(
-                &iq.id,
-                response_from,
-                response_to,
-                Some(payload),
-            )]
-        }
-        xmpp_parsers::iq::IqType::Set(query) => {
-            let term = query
-                .children()
-                .find(|child| matches!(child.name(), "nick" | "email" | "first" | "last"))
-                .map(|child| child.text())
-                .unwrap_or_default();
-            let like = format!("%{}%", term.trim());
-            let rows = match state
-                .deps
-                .app_state
-                .db_pool
-                .global_actor()
-                .clone()
-                .ask(DbQuery {
-                    sql: "SELECT username, email FROM native_users WHERE domain = ? AND (username LIKE ? OR COALESCE(email, '') LIKE ?) ORDER BY username LIMIT 50".to_string(),
-                    params: vec![domain.into(), like.clone().into(), like.into()],
-                })
-                .await
-            {
-                Ok(rows) => rows,
-                Err(error) => {
-                    warn!(error = %error, "Failed to search native users over WebSocket");
-                    return vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "wait",
-                        "internal-server-error",
-                    )];
-                }
-            };
-            let mut query = Element::builder("query", "jabber:iq:search");
-            for row in rows {
-                let username = row_value(&row, 0)
-                    .and_then(ValueExt::as_string)
-                    .unwrap_or_default();
-                if username.is_empty() {
-                    continue;
-                }
-                let email = row_value(&row, 1)
-                    .and_then(ValueExt::as_optional_string)
-                    .ok()
-                    .flatten();
-                let mut item = Element::builder("item", "jabber:iq:search")
-                    .attr("jid", format!("{username}@{domain}"))
-                    .append(
-                        Element::builder("nick", "jabber:iq:search")
-                            .append(username.clone())
-                            .build(),
-                    );
-                if let Some(email) = email {
-                    item = item.append(
-                        Element::builder("email", "jabber:iq:search")
-                            .append(email)
-                            .build(),
-                    );
-                }
-                query = query.append(item.build());
-            }
-            vec![build_iq_result_xml(
-                &iq.id,
-                response_from,
-                response_to,
-                Some(query.build()),
-            )]
-        }
-        _ => vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "modify",
-            "bad-request",
-        )],
-    }
-}
-
-async fn handle_muc_admin_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    muc_domain: &str,
-    state: &WebSocketState,
-    sender_jid: Option<&FullJid>,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-) -> Vec<String> {
-    let Some(sender_jid) = sender_jid else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "auth",
-            "not-authorized",
-        )];
-    };
-    let mut iq_with_from = iq.clone();
-    iq_with_from.from = Some(Jid::from(sender_jid.clone()));
-    let query = match parse_admin_query(&iq_with_from, muc_domain) {
-        Ok(query) => query,
-        Err(error) => return vec![build_xmpp_error_response(&iq_with_from, error)],
-    };
-    let Some(room_actor) = get_room_actor(state, &query.room_jid).await else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "cancel",
-            "item-not-found",
-        )];
-    };
-    let context = match room_actor
-        .ask(GetAdminContext {
-            sender_jid: sender_jid.clone(),
-        })
-        .await
-    {
-        Ok(context) => context,
-        Err(_) => {
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "wait",
-                "internal-server-error",
-            )]
-        }
-    };
-    let is_admin = matches!(context.affiliation, Affiliation::Owner | Affiliation::Admin)
-        || matches!(context.role, waddle_xmpp::Role::Moderator);
-    if !is_admin {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "auth",
-            "forbidden",
-        )];
-    }
-    if query.is_get {
-        let snapshot = match room_actor.ask(GetSnapshot).await {
-            Ok(snapshot) => snapshot.room,
-            Err(_) => {
-                return vec![build_iq_error_xml_with_addresses(
-                    &iq.id,
-                    response_from,
-                    response_to,
-                    "wait",
-                    "internal-server-error",
-                )]
-            }
-        };
-        let to_jid = Jid::from(sender_jid.clone());
-        if is_role_change_query(&query.items) {
-            let role_filter = query.items.iter().find_map(|item| item.role);
-            let items: Vec<(String, waddle_xmpp::Role, Option<BareJid>)> = snapshot
-                .occupants
-                .values()
-                .filter(|occupant| role_filter.is_none_or(|role| occupant.role == role))
-                .map(|occupant| {
-                    (
-                        occupant.nick.clone(),
-                        occupant.role,
-                        Some(occupant.real_jid.to_bare()),
-                    )
-                })
-                .collect();
-            return vec![iq_to_xml(build_role_result(
-                &iq.id,
-                &query.room_jid,
-                &to_jid,
-                &items,
-            ))];
-        }
-        let affiliation_filter = query.items.iter().find_map(|item| item.affiliation);
-        let items: Vec<(BareJid, Affiliation)> = if let Some(affiliation) = affiliation_filter {
-            snapshot
-                .get_jids_by_affiliation(affiliation)
-                .into_iter()
-                .map(|jid| (jid, affiliation))
-                .collect()
-        } else {
-            snapshot
-                .get_all_affiliations()
-                .into_iter()
-                .map(|entry| (entry.jid, entry.affiliation))
-                .collect()
-        };
-        return vec![iq_to_xml(build_admin_result(
-            &iq.id,
-            &query.room_jid,
-            &to_jid,
-            &items,
-        ))];
-    }
-    let room_jid = query.room_jid.clone();
-    let updates = match room_actor
-        .ask(ApplyAdminItems {
-            sender_jid: sender_jid.clone(),
-            sender_affiliation: context.affiliation,
-            sender_role: context.role,
-            items: query.items,
-        })
-        .await
-    {
-        Ok(updates) => updates,
-        Err(kameo::error::SendError::HandlerError(error)) => {
-            warn!(room = %room_jid, error = %error, "MUC admin set rejected");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "auth",
-                "forbidden",
-            )];
-        }
-        Err(error) => {
-            warn!(room = %room_jid, error = ?error, "Failed to apply MUC admin IQ");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "wait",
-                "internal-server-error",
-            )];
-        }
-    };
-    for (recipient, presence) in updates {
-        let _ = state
-            .deps
-            .protocol
-            .connection_registry
-            .send_to(&recipient, Stanza::Presence(presence))
-            .await;
-    }
-    vec![iq_to_xml(build_admin_set_result(
-        &iq.id,
-        &room_jid,
-        &Jid::from(sender_jid.clone()),
-    ))]
-}
-
 async fn handle_last_activity_iq(
     iq: &xmpp_parsers::iq::Iq,
     domain: &str,
@@ -3019,673 +2556,6 @@ async fn handle_last_activity_iq(
     )]
 }
 
-async fn handle_roster_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    domain: &str,
-    state: &WebSocketState,
-    bound_jid: Option<&FullJid>,
-    roster_interested: &mut bool,
-) -> Vec<String> {
-    let Some(full_jid) = bound_jid else {
-        return vec![build_xmpp_error_response(
-            iq,
-            XmppError::not_authorized(Some("Authenticated session required".to_string())),
-        )];
-    };
-    let user_jid = full_jid.to_bare();
-
-    if !roster_target_allowed(iq, domain, &user_jid) {
-        return vec![build_xmpp_error_response(iq, XmppError::forbidden(None))];
-    }
-
-    let storage = match roster_storage_for_state(state).await {
-        Ok(storage) => storage,
-        Err(error) => {
-            warn!(error = %error, "Failed to access roster storage");
-            return vec![build_xmpp_error_response(
-                iq,
-                XmppError::internal_server_error(None),
-            )];
-        }
-    };
-
-    if waddle_xmpp::roster::is_roster_get(iq) {
-        return handle_roster_get(iq, &storage, state, &user_jid, full_jid, roster_interested)
-            .await;
-    }
-    if waddle_xmpp::roster::is_roster_set(iq) {
-        return handle_roster_set(iq, &storage, state, &user_jid, full_jid).await;
-    }
-
-    vec![build_xmpp_error_response(iq, XmppError::bad_request(None))]
-}
-
-async fn handle_roster_get(
-    iq: &xmpp_parsers::iq::Iq,
-    storage: &DatabaseRosterStorage,
-    state: &WebSocketState,
-    user_jid: &BareJid,
-    full_jid: &FullJid,
-    roster_interested: &mut bool,
-) -> Vec<String> {
-    let query = match parse_roster_get(iq) {
-        Ok(query) => query,
-        Err(error) => return vec![build_xmpp_error_response(iq, error)],
-    };
-
-    // Atomic snapshot: items + version under the same per-user lock so the
-    // returned ver always identifies the returned roster state (XEP-0237 §2.6).
-    let (rows, version) = match storage.snapshot_roster(user_jid).await {
-        Ok(pair) => pair,
-        Err(error) => {
-            warn!(user = %user_jid, error = %error, "Failed to snapshot roster");
-            return vec![build_xmpp_error_response(
-                iq,
-                XmppError::internal_server_error(None),
-            )];
-        }
-    };
-    let items = match roster_rows_to_items(rows) {
-        Ok(items) => items,
-        Err(error) => {
-            warn!(user = %user_jid, error = %error, "Failed to convert roster rows");
-            return vec![build_xmpp_error_response(
-                iq,
-                XmppError::internal_server_error(None),
-            )];
-        }
-    };
-
-    state
-        .deps
-        .protocol
-        .connection_registry
-        .mark_roster_interested(full_jid);
-    *roster_interested = true;
-    // XEP-0237 §2.6: matching ver -> empty <iq type='result'/> (no <query>);
-    // anything else (Absent, Bootstrap, Cached(stale)) -> full roster + ver.
-    if let Some(cached) = query.ver.cached() {
-        if cached == &version {
-            return vec![iq_to_xml(build_roster_result_empty(iq))];
-        }
-    }
-    vec![iq_to_xml(build_roster_result(iq, &items, Some(&version)))]
-}
-
-async fn handle_roster_set(
-    iq: &xmpp_parsers::iq::Iq,
-    storage: &DatabaseRosterStorage,
-    state: &WebSocketState,
-    user_jid: &BareJid,
-    full_jid: &FullJid,
-) -> Vec<String> {
-    let query = match parse_roster_set(iq) {
-        Ok(query) => query,
-        Err(error) => return vec![build_xmpp_error_response(iq, error)],
-    };
-    let requested = query
-        .items
-        .first()
-        .expect("parse_roster_set guarantees one item");
-
-    let mut removed_item = None;
-    // Hold the per-user mutation lock from the start of the storage write
-    // through the end of push fanout below (XEP-0237 §2.6 — pushes for
-    // mutation N must be enqueued before mutation N+1's pushes can race onto
-    // the recipient socket). The guard is dropped at the end of this block.
-    let (set_result, version, _user_lock) = if requested.subscription.is_remove() {
-        // Snapshot the item before delete so we can run subscription side effects.
-        removed_item = match storage.get_roster_item(user_jid, &requested.jid).await {
-            Ok(Some(row)) => match roster_row_to_item(row) {
-                Ok(item) => Some(item),
-                Err(error) => {
-                    warn!(user = %user_jid, contact = %requested.jid, error = %error, "Failed to convert removed roster item");
-                    return vec![build_xmpp_error_response(
-                        iq,
-                        XmppError::internal_server_error(None),
-                    )];
-                }
-            },
-            Ok(None) => None,
-            Err(error) => {
-                warn!(user = %user_jid, contact = %requested.jid, error = %error, "Failed to load roster item before remove");
-                return vec![build_xmpp_error_response(
-                    iq,
-                    XmppError::internal_server_error(None),
-                )];
-            }
-        };
-        match storage
-            .apply_roster_change(user_jid, RosterRowChange::Remove(requested.jid.clone()))
-            .await
-        {
-            Ok((mutation, lock)) => (
-                RosterSetResult::Removed(requested.jid.clone()),
-                mutation.version,
-                lock,
-            ),
-            Err(RosterStorageError::ItemNotFound) => {
-                return vec![build_xmpp_error_response(
-                    iq,
-                    XmppError::item_not_found(None),
-                )];
-            }
-            Err(error) => {
-                warn!(user = %user_jid, contact = %requested.jid, error = %error, "Failed to remove roster item");
-                return vec![build_xmpp_error_response(
-                    iq,
-                    XmppError::internal_server_error(None),
-                )];
-            }
-        }
-    } else {
-        let mut item = match storage.get_roster_item(user_jid, &requested.jid).await {
-            Ok(Some(row)) => match roster_row_to_item(row) {
-                Ok(item) => item,
-                Err(error) => {
-                    warn!(user = %user_jid, contact = %requested.jid, error = %error, "Failed to convert roster item");
-                    return vec![build_xmpp_error_response(
-                        iq,
-                        XmppError::internal_server_error(None),
-                    )];
-                }
-            },
-            Ok(None) => RosterItem::new(requested.jid.clone()),
-            Err(error) => {
-                warn!(user = %user_jid, contact = %requested.jid, error = %error, "Failed to load roster item");
-                return vec![build_xmpp_error_response(
-                    iq,
-                    XmppError::internal_server_error(None),
-                )];
-            }
-        };
-
-        item.name = requested.name.clone();
-        item.groups = requested.groups.clone();
-
-        match storage
-            .apply_roster_change(user_jid, RosterRowChange::Upsert(roster_item_to_row(&item)))
-            .await
-        {
-            Ok((mutation, lock)) => {
-                let result = match mutation.kind {
-                    RosterRowMutationKind::Added(_) => RosterSetResult::Added(item),
-                    RosterRowMutationKind::Updated(_) => RosterSetResult::Updated(item),
-                    RosterRowMutationKind::Removed(_) => {
-                        unreachable!("Upsert never reports Removed")
-                    }
-                };
-                (result, mutation.version, lock)
-            }
-            Err(error) => {
-                warn!(user = %user_jid, contact = %item.jid, error = %error, "Failed to store roster item");
-                return vec![build_xmpp_error_response(
-                    iq,
-                    XmppError::internal_server_error(None),
-                )];
-            }
-        }
-    };
-
-    let push_item = set_result.to_push_item();
-    let mut frames = Vec::new();
-    if state
-        .deps
-        .protocol
-        .connection_registry
-        .is_roster_interested(full_jid)
-    {
-        frames.push(iq_to_xml(build_roster_push(
-            &format!("roster-push-{}", uuid::Uuid::new_v4()),
-            user_jid,
-            full_jid,
-            &push_item,
-            Some(&version),
-        )));
-    }
-
-    send_roster_push_to_sibling_resources(state, user_jid, full_jid, &push_item, &version).await;
-    // Drop the user_jid mutation lock before invoking subscription side
-    // effects on the *contact's* roster — the side-effect path acquires the
-    // contact's lock, and holding two user-locks simultaneously could
-    // deadlock against a concurrent flow that touches the same pair in the
-    // opposite role (PR #336 review). The user_jid pushes have all been
-    // enqueued by this point, so XEP-0237 §2.6 ordering for user_jid is
-    // already preserved.
-    drop(_user_lock);
-    if let Some(item) = removed_item.as_ref() {
-        send_roster_remove_subscription_side_effects(state, storage, user_jid, item).await;
-    }
-    frames.push(iq_to_xml(build_roster_result_empty(iq)));
-    frames
-}
-
-async fn send_roster_remove_subscription_side_effects(
-    state: &WebSocketState,
-    storage: &DatabaseRosterStorage,
-    user_jid: &BareJid,
-    removed_item: &RosterItem,
-) {
-    if matches!(
-        removed_item.subscription,
-        Subscription::To | Subscription::Both
-    ) {
-        send_roster_remove_subscription_stanza(
-            state,
-            storage,
-            user_jid,
-            &removed_item.jid,
-            SubscriptionType::Unsubscribe,
-        )
-        .await;
-    }
-    if matches!(
-        removed_item.subscription,
-        Subscription::From | Subscription::Both
-    ) {
-        send_roster_remove_subscription_stanza(
-            state,
-            storage,
-            user_jid,
-            &removed_item.jid,
-            SubscriptionType::Unsubscribed,
-        )
-        .await;
-    }
-}
-
-async fn send_roster_remove_subscription_stanza(
-    state: &WebSocketState,
-    storage: &DatabaseRosterStorage,
-    from: &BareJid,
-    to: &BareJid,
-    subscription_type: SubscriptionType,
-) {
-    if let Ok(Some(row)) = storage.get_roster_item(to, from).await {
-        match roster_row_to_item(row) {
-            Ok(mut item) => {
-                match subscription_type {
-                    SubscriptionType::Unsubscribe => {
-                        SubscriptionStateMachine::apply_outbound_unsubscribed(&mut item);
-                    }
-                    SubscriptionType::Unsubscribed => {
-                        SubscriptionStateMachine::apply_inbound_unsubscribed(&mut item);
-                    }
-                    SubscriptionType::Subscribe | SubscriptionType::Subscribed => {}
-                }
-                match storage
-                    .apply_roster_change(to, RosterRowChange::Upsert(roster_item_to_row(&item)))
-                    .await
-                {
-                    Ok((mutation, _lock)) => {
-                        // _lock held until the push enqueue below completes.
-                        send_roster_push_to_all_resources(state, to, &item, &mutation.version)
-                            .await;
-                    }
-                    Err(error) => {
-                        warn!(error = %error, user = %to, contact = %from, "Failed to update contact roster after removal side effect");
-                    }
-                }
-            }
-            Err(error) => {
-                warn!(error = %error, user = %to, contact = %from, "Failed to convert contact roster item after removal side effect");
-            }
-        }
-    }
-
-    let stanza = Stanza::Presence(build_subscription_presence(
-        subscription_type,
-        from,
-        to,
-        None,
-        &[],
-    ));
-    let live_resources = state
-        .deps
-        .protocol
-        .connection_registry
-        .get_roster_interested_resources_for_user(to);
-    let mut delivered_resources = Vec::new();
-    for resource in &live_resources {
-        if state
-            .deps
-            .protocol
-            .connection_registry
-            .try_send_to(resource, stanza.clone())
-            == BroadcastOutcome::Delivered
-        {
-            delivered_resources.push(resource.clone());
-        }
-    }
-
-    let detached = match state
-        .deps
-        .protocol
-        .sm_session_registry
-        .interested_detached_resources_for_user(to)
-        .await
-    {
-        Ok(resources) => resources,
-        Err(error) => {
-            warn!(error = %error, user = %to, "Failed to list detached interested resources for roster removal side effect");
-            return;
-        }
-    };
-    let detached: Vec<_> = detached
-        .into_iter()
-        .filter(|resource| !delivered_resources.contains(resource))
-        .collect();
-    for resource in detached {
-        match state
-            .deps
-            .protocol
-            .sm_session_registry
-            .record_stanza_for_detached_resource(&resource, &stanza)
-            .await
-        {
-            Ok(true) => {}
-            Ok(false) => {
-                let _ = state
-                    .deps
-                    .protocol
-                    .connection_registry
-                    .try_send_to(&resource, stanza.clone());
-            }
-            Err(error) => {
-                warn!(error = %error, resource = %resource, "Failed to record roster removal subscription side effect");
-            }
-        }
-    }
-}
-
-async fn send_roster_push_to_all_resources(
-    state: &WebSocketState,
-    user_jid: &BareJid,
-    item: &RosterItem,
-    version: &RosterVersion,
-) {
-    let live_resources = state
-        .deps
-        .protocol
-        .connection_registry
-        .get_roster_interested_resources_for_user(user_jid);
-    let mut delivered_resources = Vec::new();
-    for resource in &live_resources {
-        let push = build_roster_push(
-            &format!("roster-push-{}", uuid::Uuid::new_v4()),
-            user_jid,
-            resource,
-            item,
-            Some(version),
-        );
-        if state
-            .deps
-            .protocol
-            .connection_registry
-            .try_send_to(resource, Stanza::Iq(push))
-            == BroadcastOutcome::Delivered
-        {
-            delivered_resources.push(resource.clone());
-        }
-    }
-    let detached = match state
-        .deps
-        .protocol
-        .sm_session_registry
-        .interested_detached_resources_for_user(user_jid)
-        .await
-    {
-        Ok(resources) => resources,
-        Err(error) => {
-            warn!(error = %error, user = %user_jid, "Failed to list detached resources for roster push");
-            return;
-        }
-    };
-    let detached: Vec<_> = detached
-        .into_iter()
-        .filter(|resource| !delivered_resources.contains(resource))
-        .collect();
-    for resource in detached {
-        let push = build_roster_push(
-            &format!("roster-push-{}", uuid::Uuid::new_v4()),
-            user_jid,
-            &resource,
-            item,
-            Some(version),
-        );
-        let stanza = Stanza::Iq(push);
-        match state
-            .deps
-            .protocol
-            .sm_session_registry
-            .record_stanza_for_detached_resource(&resource, &stanza)
-            .await
-        {
-            Ok(true) => delivered_resources.push(resource.clone()),
-            Ok(false) => {
-                if state
-                    .deps
-                    .protocol
-                    .connection_registry
-                    .try_send_to(&resource, stanza)
-                    == BroadcastOutcome::Delivered
-                {
-                    delivered_resources.push(resource.clone());
-                }
-            }
-            Err(error) => {
-                warn!(error = %error, resource = %resource, "Failed to record detached roster push");
-            }
-        }
-    }
-    for resource in state
-        .deps
-        .protocol
-        .connection_registry
-        .get_roster_interested_resources_for_user(user_jid)
-        .into_iter()
-        .filter(|resource| !delivered_resources.contains(resource))
-    {
-        let push = build_roster_push(
-            &format!("roster-push-{}", uuid::Uuid::new_v4()),
-            user_jid,
-            &resource,
-            item,
-            Some(version),
-        );
-        let _ = state
-            .deps
-            .protocol
-            .connection_registry
-            .try_send_to(&resource, Stanza::Iq(push));
-    }
-}
-
-async fn send_roster_push_to_sibling_resources(
-    state: &WebSocketState,
-    user_jid: &BareJid,
-    source_jid: &FullJid,
-    item: &RosterItem,
-    version: &RosterVersion,
-) {
-    let live_resources: Vec<_> = state
-        .deps
-        .protocol
-        .connection_registry
-        .get_roster_interested_resources_for_user(user_jid)
-        .into_iter()
-        .filter(|resource| resource != source_jid)
-        .collect();
-    let mut delivered_resources = Vec::new();
-    for resource in &live_resources {
-        let push = build_roster_push(
-            &format!("roster-push-{}", uuid::Uuid::new_v4()),
-            user_jid,
-            resource,
-            item,
-            Some(version),
-        );
-        if state
-            .deps
-            .protocol
-            .connection_registry
-            .try_send_to(resource, Stanza::Iq(push))
-            == BroadcastOutcome::Delivered
-        {
-            delivered_resources.push(resource.clone());
-        }
-    }
-    let detached = match state
-        .deps
-        .protocol
-        .sm_session_registry
-        .interested_detached_resources_for_user(user_jid)
-        .await
-    {
-        Ok(resources) => resources,
-        Err(error) => {
-            warn!(error = %error, user = %user_jid, "Failed to list detached roster-interested resources");
-            return;
-        }
-    };
-    let detached: Vec<_> = detached
-        .into_iter()
-        .filter(|resource| resource != source_jid)
-        .filter(|resource| !delivered_resources.contains(resource))
-        .collect();
-    for resource in detached {
-        let push = build_roster_push(
-            &format!("roster-push-{}", uuid::Uuid::new_v4()),
-            user_jid,
-            &resource,
-            item,
-            Some(version),
-        );
-        let stanza = Stanza::Iq(push.clone());
-        match state
-            .deps
-            .protocol
-            .sm_session_registry
-            .record_stanza_for_detached_resource(&resource, &stanza)
-            .await
-        {
-            Ok(true) => delivered_resources.push(resource.clone()),
-            Ok(false) => {
-                let is_interested = state
-                    .deps
-                    .protocol
-                    .connection_registry
-                    .is_roster_interested(&resource);
-                if is_interested
-                    && state
-                        .deps
-                        .protocol
-                        .connection_registry
-                        .try_send_to(&resource, stanza)
-                        == BroadcastOutcome::Delivered
-                {
-                    delivered_resources.push(resource.clone());
-                }
-            }
-            Err(error) => {
-                warn!(error = %error, resource = %resource, "Failed to record detached roster push");
-            }
-        }
-    }
-    for resource in state
-        .deps
-        .protocol
-        .connection_registry
-        .get_roster_interested_resources_for_user(user_jid)
-        .into_iter()
-        .filter(|resource| resource != source_jid)
-        .filter(|resource| !delivered_resources.contains(resource))
-    {
-        let push = build_roster_push(
-            &format!("roster-push-{}", uuid::Uuid::new_v4()),
-            user_jid,
-            &resource,
-            item,
-            Some(version),
-        );
-        let _ = state
-            .deps
-            .protocol
-            .connection_registry
-            .try_send_to(&resource, Stanza::Iq(push));
-    }
-}
-
-fn roster_target_allowed(iq: &xmpp_parsers::iq::Iq, domain: &str, user_jid: &BareJid) -> bool {
-    match iq.to.as_ref() {
-        None => true,
-        Some(to) if to.to_bare() == *user_jid => true,
-        Some(to) if to.resource().is_none() && to.to_bare().as_str() == domain => true,
-        _ => false,
-    }
-}
-
-#[derive(Debug, thiserror::Error)]
-enum RosterConvertError {
-    #[error("invalid stored roster jid '{jid}': {error}")]
-    Jid { jid: String, error: String },
-    #[error("invalid stored roster subscription: {0}")]
-    Subscription(String),
-    #[error("invalid stored roster ask: {0}")]
-    Ask(String),
-}
-
-async fn roster_storage_for_state(
-    state: &WebSocketState,
-) -> Result<DatabaseRosterStorage, RosterStorageError> {
-    let db = global_database(state).await?;
-    Ok(DatabaseRosterStorage::new(db))
-}
-
-fn roster_rows_to_items(rows: Vec<RosterItemRow>) -> Result<Vec<RosterItem>, RosterConvertError> {
-    rows.into_iter().map(roster_row_to_item).collect()
-}
-
-fn roster_row_to_item(row: RosterItemRow) -> Result<RosterItem, RosterConvertError> {
-    let jid = row
-        .contact_jid
-        .parse::<BareJid>()
-        .map_err(|error: jid::Error| RosterConvertError::Jid {
-            jid: row.contact_jid.clone(),
-            error: error.to_string(),
-        })?;
-    let subscription = row
-        .subscription
-        .parse::<Subscription>()
-        .map_err(|error| RosterConvertError::Subscription(error.to_string()))?;
-    let ask = row
-        .ask
-        .as_deref()
-        .map(str::parse::<AskType>)
-        .transpose()
-        .map_err(|error| RosterConvertError::Ask(error.to_string()))?;
-
-    Ok(RosterItem {
-        jid,
-        name: row.name,
-        subscription,
-        ask,
-        approved: row.approved,
-        groups: row.groups,
-    })
-}
-
-fn roster_item_to_row(item: &RosterItem) -> RosterItemRow {
-    RosterItemRow {
-        contact_jid: item.jid.to_string(),
-        name: item.name.clone(),
-        subscription: item.subscription.as_str().to_string(),
-        ask: item.ask.map(|ask| ask.as_str().to_string()),
-        approved: item.approved,
-        groups: item.groups.clone(),
-    }
-}
-
 fn build_xmpp_error_response(request_iq: &xmpp_parsers::iq::Iq, err: XmppError) -> String {
     match err {
         XmppError::Stanza {
@@ -3729,718 +2599,6 @@ async fn global_database(state: &WebSocketState) -> Result<Database, RosterStora
         .ask(GetDatabase)
         .await
         .map_err(|error| RosterStorageError::ConnectionFailed(error.to_string()))
-}
-
-async fn handle_vcard_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    sender_jid: Option<&FullJid>,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-) -> Vec<String> {
-    let Some(sender_jid) = sender_jid else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "auth",
-            "not-authorized",
-        )];
-    };
-
-    let db = match global_database(state).await {
-        Ok(db) => Arc::new(db),
-        Err(error) => {
-            warn!(error = %error, "Failed to access database for vCard IQ");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "wait",
-                "internal-server-error",
-            )];
-        }
-    };
-    let store = VCardStore::new(Arc::clone(&db));
-
-    if waddle_xmpp::xep::xep0054::is_vcard_get(iq) {
-        let target_jid = iq
-            .to
-            .as_ref()
-            .map(|jid| jid.to_bare())
-            .unwrap_or_else(|| sender_jid.to_bare());
-        let response = match store.get(&target_jid).await {
-            Ok(Some(xml)) => match xml.parse::<Element>() {
-                Ok(vcard) => xmpp_parsers::iq::Iq {
-                    from: iq.to.clone(),
-                    to: iq.from.clone(),
-                    id: iq.id.clone(),
-                    payload: xmpp_parsers::iq::IqType::Result(Some(vcard)),
-                },
-                Err(error) => {
-                    warn!(target = %target_jid, error = %error, "Stored vCard XML is invalid");
-                    return vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "wait",
-                        "internal-server-error",
-                    )];
-                }
-            },
-            Ok(None) => match avatar_vcard_from_user_profile(Arc::clone(&db), &target_jid).await {
-                Ok(Some(vcard)) => waddle_xmpp::xep::xep0054::build_vcard_response(iq, &vcard),
-                Ok(None) => {
-                    return vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "cancel",
-                        "item-not-found",
-                    )];
-                }
-                Err(error) => {
-                    warn!(target = %target_jid, error = %error, "Failed to load avatar vCard fallback");
-                    return vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "wait",
-                        "internal-server-error",
-                    )];
-                }
-            },
-            Err(error) => {
-                warn!(target = %target_jid, error = %error, "Failed to load vCard");
-                return vec![build_iq_error_xml_with_addresses(
-                    &iq.id,
-                    response_from,
-                    response_to,
-                    "wait",
-                    "internal-server-error",
-                )];
-            }
-        };
-        return vec![iq_to_xml(response)];
-    }
-
-    if let Some(to) = &iq.to {
-        if to.to_bare() != sender_jid.to_bare() {
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "auth",
-                "forbidden",
-            )];
-        }
-    }
-
-    let xmpp_parsers::iq::IqType::Set(vcard) = &iq.payload else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "modify",
-            "bad-request",
-        )];
-    };
-
-    if let Err(error) = store.set(&sender_jid.to_bare(), &String::from(vcard)).await {
-        warn!(jid = %sender_jid, error = %error, "Failed to store vCard");
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "wait",
-            "internal-server-error",
-        )];
-    }
-
-    vec![iq_to_xml(waddle_xmpp::xep::xep0054::build_vcard_success(
-        iq,
-    ))]
-}
-
-async fn handle_private_storage_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    sender_jid: Option<&FullJid>,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-) -> Vec<String> {
-    let Some(sender_jid) = sender_jid else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "auth",
-            "not-authorized",
-        )];
-    };
-    let bare_jid = sender_jid.to_bare();
-
-    if let Some(key) = waddle_xmpp::xep::xep0049::parse_private_storage_get(iq) {
-        if waddle_xmpp::xep::xep0048::is_legacy_bookmarks_namespace(&key.namespace) {
-            let stored_items = state
-                .deps
-                .protocol
-                .pubsub_storage
-                .get_items(&bare_jid, waddle_xmpp::xep::xep0402::PEP_NODE, None, &[])
-                .await
-                .unwrap_or_default();
-            let native_bookmarks: Vec<waddle_xmpp::xep::xep0402::Bookmark> = stored_items
-                .iter()
-                .filter_map(|item| {
-                    let xml = item.payload_xml.as_ref()?;
-                    let elem: Element = xml.parse().ok()?;
-                    waddle_xmpp::xep::xep0402::parse_bookmark(&item.id, &elem).ok()
-                })
-                .collect();
-            let legacy: Vec<waddle_xmpp::xep::xep0048::LegacyBookmark> = native_bookmarks
-                .iter()
-                .map(waddle_xmpp::xep::xep0048::from_native_bookmark)
-                .collect();
-            let bookmarks = waddle_xmpp::xep::xep0048::build_legacy_bookmarks_element(&legacy);
-            let response = waddle_xmpp::xep::xep0049::build_private_storage_result(
-                iq,
-                Some(&String::from(&bookmarks)),
-                &key,
-            );
-            return vec![iq_to_xml(response)];
-        }
-
-        let row = match state
-            .deps
-            .app_state
-            .db_pool
-            .global_actor()
-            .clone()
-            .ask(DbQueryOne {
-                sql: "SELECT xml_content FROM private_xml_storage WHERE jid = ? AND namespace = ?"
-                    .to_string(),
-                params: vec![
-                    Value::from(bare_jid.to_string()),
-                    Value::from(key.namespace.clone()),
-                ],
-            })
-            .await
-        {
-            Ok(row) => row,
-            Err(error) => {
-                warn!(jid = %bare_jid, namespace = %key.namespace, error = %error, "Failed to load private XML");
-                return vec![build_iq_error_xml_with_addresses(
-                    &iq.id,
-                    response_from,
-                    response_to,
-                    "wait",
-                    "internal-server-error",
-                )];
-            }
-        };
-        let stored = match row {
-            Some(values) => match row_value(&values, 0).and_then(|value| value.as_string()) {
-                Ok(value) => Some(value),
-                Err(error) => {
-                    warn!(jid = %bare_jid, namespace = %key.namespace, error = %error, "Failed to decode private XML row");
-                    return vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "wait",
-                        "internal-server-error",
-                    )];
-                }
-            },
-            None => None,
-        };
-        let response =
-            waddle_xmpp::xep::xep0049::build_private_storage_result(iq, stored.as_deref(), &key);
-        return vec![iq_to_xml(response)];
-    }
-
-    if let Some((key, xml_content)) = waddle_xmpp::xep::xep0049::parse_private_storage_set(iq) {
-        if waddle_xmpp::xep::xep0048::is_legacy_bookmarks_namespace(&key.namespace) {
-            if let Ok(elem) = xml_content.parse::<Element>() {
-                for legacy in waddle_xmpp::xep::xep0048::parse_legacy_bookmarks(&elem) {
-                    if let Some(native) = waddle_xmpp::xep::xep0048::to_native_bookmark(&legacy) {
-                        let bookmark_elem =
-                            waddle_xmpp::xep::xep0402::build_bookmark_element(&native);
-                        let item =
-                            PubSubItem::new(Some(native.jid.to_string()), Some(bookmark_elem));
-                        let _ = state
-                            .deps
-                            .protocol
-                            .pubsub_storage
-                            .publish_item(
-                                &bare_jid,
-                                waddle_xmpp::xep::xep0402::PEP_NODE,
-                                &item,
-                                Some(&bare_jid),
-                                true,
-                            )
-                            .await;
-                    }
-                }
-            }
-            return vec![iq_to_xml(
-                waddle_xmpp::xep::xep0049::build_private_storage_success(iq),
-            )];
-        }
-
-        if let Err(error) = state
-            .deps
-            .app_state
-            .db_pool
-            .global_actor()
-            .clone()
-            .ask(DbExecute {
-                sql: "INSERT OR REPLACE INTO private_xml_storage (jid, namespace, xml_content, updated_at) VALUES (?, ?, ?, datetime('now'))".to_string(),
-                params: vec![
-                    Value::from(bare_jid.to_string()),
-                    Value::from(key.namespace),
-                    Value::from(xml_content),
-                ],
-            })
-            .await
-        {
-            warn!(jid = %bare_jid, error = %error, "Failed to store private XML");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "wait",
-                "internal-server-error",
-            )];
-        }
-        return vec![iq_to_xml(
-            waddle_xmpp::xep::xep0049::build_private_storage_success(iq),
-        )];
-    }
-
-    vec![build_iq_error_xml_with_addresses(
-        &iq.id,
-        response_from,
-        response_to,
-        "modify",
-        "bad-request",
-    )]
-}
-
-async fn avatar_vcard_from_user_profile(
-    db: Arc<Database>,
-    target_jid: &BareJid,
-) -> Result<Option<VCard>, String> {
-    let Some(localpart) = target_jid.node().map(|node| node.to_string()) else {
-        return Ok(None);
-    };
-
-    let conn = db.guard().await.map_err(|error| error.to_string())?;
-    let mut rows = conn
-        .query(
-            "SELECT avatar_url FROM users WHERE xmpp_localpart = ? LIMIT 1",
-            [localpart.as_str()],
-        )
-        .await
-        .map_err(|error| error.to_string())?;
-    let Some(row) = rows.next().await.map_err(|error| error.to_string())? else {
-        return Ok(None);
-    };
-    let avatar_url: Option<String> = row.get(0).map_err(|error| error.to_string())?;
-    let Some(avatar_url) = avatar_url.filter(|url| url.starts_with("https://")) else {
-        return Ok(None);
-    };
-
-    let vcard = VCard {
-        photo: Some(VCardPhoto::External { url: avatar_url }),
-        ..Default::default()
-    };
-
-    Ok(Some(vcard))
-}
-
-async fn handle_blocking_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    sender_jid: Option<&FullJid>,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-    state_machine: Option<&mut waddle_xmpp::protocol::XmppStateMachine>,
-) -> Vec<String> {
-    let Some(sender_jid) = sender_jid else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "auth",
-            "not-authorized",
-        )];
-    };
-    let user_bare = sender_jid.to_bare();
-    let db = match global_database(state).await {
-        Ok(db) => db,
-        Err(error) => {
-            warn!(error = %error, "Failed to access database for blocking IQ");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "wait",
-                "internal-server-error",
-            )];
-        }
-    };
-    let storage = DatabaseBlockingStorage::new(db);
-    let request = match waddle_xmpp::xep::xep0191::parse_blocking_request(iq) {
-        Ok(request) => request,
-        Err(error) => {
-            warn!(error = %error, "Invalid blocking IQ");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "modify",
-                "bad-request",
-            )];
-        }
-    };
-
-    let response = match request {
-        waddle_xmpp::xep::xep0191::BlockingRequest::GetBlocklist => {
-            return match storage.get_blocklist(&user_bare).await {
-                Ok(blocked) => vec![iq_to_xml(
-                    waddle_xmpp::xep::xep0191::build_blocklist_response(iq, &blocked),
-                )],
-                Err(error) => {
-                    warn!(jid = %user_bare, error = %error, "Failed to load blocklist");
-                    vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "wait",
-                        "internal-server-error",
-                    )]
-                }
-            };
-        }
-        waddle_xmpp::xep::xep0191::BlockingRequest::Block(jids) => {
-            if let Err(error) = storage.add_blocks(&user_bare, &jids).await {
-                warn!(jid = %user_bare, error = %error, "Failed to add blocks");
-                return vec![build_iq_error_xml_with_addresses(
-                    &iq.id,
-                    response_from,
-                    response_to,
-                    "wait",
-                    "internal-server-error",
-                )];
-            }
-            send_blocking_presence_side_effects(state, &user_bare, &jids, true).await;
-            send_blocking_pushes(state, &user_bare, true, &jids).await;
-            vec![iq_to_xml(
-                waddle_xmpp::xep::xep0191::build_blocking_success(iq),
-            )]
-        }
-        waddle_xmpp::xep::xep0191::BlockingRequest::Unblock(jids) => {
-            let unblocked_jids = if jids.is_empty() {
-                let current = match storage.get_blocklist(&user_bare).await {
-                    Ok(current) => current,
-                    Err(error) => {
-                        warn!(jid = %user_bare, error = %error, "Failed to load blocklist before unblock-all");
-                        return vec![build_iq_error_xml_with_addresses(
-                            &iq.id,
-                            response_from,
-                            response_to,
-                            "wait",
-                            "internal-server-error",
-                        )];
-                    }
-                };
-                if let Err(error) = storage.remove_all_blocks(&user_bare).await {
-                    warn!(jid = %user_bare, error = %error, "Failed to remove all blocks");
-                    return vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "wait",
-                        "internal-server-error",
-                    )];
-                }
-                current
-            } else {
-                if let Err(error) = storage.remove_blocks(&user_bare, &jids).await {
-                    warn!(jid = %user_bare, error = %error, "Failed to remove blocks");
-                    return vec![build_iq_error_xml_with_addresses(
-                        &iq.id,
-                        response_from,
-                        response_to,
-                        "wait",
-                        "internal-server-error",
-                    )];
-                }
-                jids
-            };
-            send_blocking_presence_side_effects(state, &user_bare, &unblocked_jids, false).await;
-            send_blocking_pushes(state, &user_bare, false, &unblocked_jids).await;
-            vec![iq_to_xml(
-                waddle_xmpp::xep::xep0191::build_blocking_success(iq),
-            )]
-        }
-    };
-
-    // After a successful Block/Unblock storage mutation, mirror the
-    // updated XEP-0191 list into the per-connection
-    // [`waddle_xmpp::protocol::XmppStateMachine`] so the dispatcher's
-    // session-state snapshot reflects the change for subsequent
-    // sender / recipient passes. Without this, blocks added live on
-    // a session would not take effect until the next bind (PR13's
-    // load-at-bind seed). Failing to reload the storage view is
-    // logged at WARN and leaves the SM blocklist unchanged: the
-    // storage layer is authoritative on disk and the next bind will
-    // reload, while the request itself already succeeded for the
-    // client.
-    if let Some(sm) = state_machine {
-        match storage.list_blocked_jids(&user_bare).await {
-            Ok(jids) => {
-                sm.set_blocklist(waddle_xmpp::protocol::Blocklist::new(jids));
-            }
-            Err(error) => {
-                warn!(
-                    jid = %user_bare,
-                    %error,
-                    "Failed to refresh in-memory blocklist after XEP-0191 IQ-set; \
-                     dispatcher snapshot will catch up on next bind"
-                );
-            }
-        }
-    }
-
-    response
-}
-
-async fn send_blocking_presence_side_effects(
-    state: &WebSocketState,
-    user_bare: &BareJid,
-    jids: &[String],
-    blocked: bool,
-) {
-    let storage = match roster_storage_for_state(state).await {
-        Ok(storage) => storage,
-        Err(error) => {
-            warn!(jid = %user_bare, error = %error, "Failed to access roster storage for XEP-0191 presence side effects");
-            return;
-        }
-    };
-    let subscribers = match storage.get_presence_subscribers(user_bare).await {
-        Ok(subscribers) => subscribers,
-        Err(error) => {
-            warn!(jid = %user_bare, error = %error, "Failed to load presence subscribers for XEP-0191 presence side effects");
-            return;
-        }
-    };
-
-    let subscriber_bares: HashSet<BareJid> = subscribers
-        .into_iter()
-        .filter_map(|jid| match jid.parse::<Jid>() {
-            Ok(jid) => Some(jid.to_bare()),
-            Err(error) => {
-                warn!(jid, %error, "Skipping invalid stored roster subscriber JID");
-                None
-            }
-        })
-        .collect();
-
-    let mut targets = Vec::new();
-    let mut seen = HashSet::new();
-    for jid in jids {
-        let Ok(target) = jid.parse::<Jid>() else {
-            warn!(
-                jid,
-                "Skipping invalid XEP-0191 target JID for presence side effects"
-            );
-            continue;
-        };
-        let target_bare = target.to_bare();
-        if subscriber_bares.contains(&target_bare) && seen.insert(target_bare.clone()) {
-            targets.push(target_bare);
-        }
-    }
-
-    for target in targets {
-        if blocked {
-            send_unavailable_presence_from_user_to_user(state, user_bare, &target).await;
-        } else {
-            send_current_presence_from_user_to_user(state, user_bare, &target).await;
-        }
-    }
-}
-
-async fn send_blocking_pushes(
-    state: &WebSocketState,
-    user_bare: &BareJid,
-    blocked: bool,
-    jids: &[String],
-) {
-    for resource_jid in state
-        .deps
-        .protocol
-        .connection_registry
-        .get_resources_for_user(user_bare)
-    {
-        let push = if blocked {
-            waddle_xmpp::xep::xep0191::build_block_push(&resource_jid.clone().into(), jids)
-        } else {
-            waddle_xmpp::xep::xep0191::build_unblock_push(&resource_jid.clone().into(), jids)
-        };
-        let _ = state
-            .deps
-            .protocol
-            .connection_registry
-            .send_to(&resource_jid, Stanza::Iq(push))
-            .await;
-    }
-}
-
-async fn handle_push_iq(
-    iq: &xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    sender_jid: Option<&FullJid>,
-    response_from: Option<&str>,
-    response_to: Option<&str>,
-) -> Vec<String> {
-    let Some(sender_jid) = sender_jid else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "auth",
-            "not-authorized",
-        )];
-    };
-    let bare_jid = sender_jid.to_bare().to_string();
-
-    if is_push_enable(iq) {
-        let Some(enable) = parse_push_enable(iq) else {
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "modify",
-                "bad-request",
-            )];
-        };
-        let endpoint = enable
-            .options
-            .iter()
-            .find(|(key, _)| key == "endpoint")
-            .map(|(_, value)| value.clone());
-        let p256dh = enable
-            .options
-            .iter()
-            .find(|(key, _)| key == "p256dh")
-            .map(|(_, value)| value.clone());
-        let auth_key = enable
-            .options
-            .iter()
-            .find(|(key, _)| key == "auth")
-            .map(|(_, value)| value.clone());
-
-        if !endpoint.as_deref().is_some_and(valid_push_endpoint)
-            || p256dh.is_none()
-            || auth_key.is_none()
-        {
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "modify",
-                "bad-request",
-            )];
-        }
-
-        let subscription = waddle_xmpp::push::PushSubscription {
-            user_jid: bare_jid.clone(),
-            service_jid: enable.jid,
-            node: enable.node,
-            endpoint,
-            p256dh,
-            auth_key,
-        };
-        if let Err(error) = state.deps.protocol.push_store.register(subscription).await {
-            warn!(user = %bare_jid, error = %error, "Failed to register push subscription");
-            return vec![build_iq_error_xml_with_addresses(
-                &iq.id,
-                response_from,
-                response_to,
-                "wait",
-                "internal-server-error",
-            )];
-        }
-        return vec![iq_to_xml(build_push_enable_result(iq))];
-    }
-
-    let Some(disable) = parse_push_disable(iq) else {
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "modify",
-            "bad-request",
-        )];
-    };
-    if let Err(error) = state
-        .deps
-        .protocol
-        .push_store
-        .remove(&bare_jid, &disable.jid, disable.node.as_deref())
-        .await
-    {
-        warn!(user = %bare_jid, error = %error, "Failed to remove push subscription");
-        return vec![build_iq_error_xml_with_addresses(
-            &iq.id,
-            response_from,
-            response_to,
-            "wait",
-            "internal-server-error",
-        )];
-    }
-    vec![iq_to_xml(build_push_disable_result(iq))]
-}
-
-fn valid_push_endpoint(value: &str) -> bool {
-    let Ok(url) = Url::parse(value) else {
-        return false;
-    };
-    if url.scheme() != "https" {
-        return false;
-    }
-    match url.host() {
-        Some(Host::Ipv4(addr)) => {
-            !(addr.is_private()
-                || addr.is_loopback()
-                || addr.is_link_local()
-                || addr.is_multicast()
-                || addr.is_broadcast()
-                || addr.is_unspecified())
-        }
-        Some(Host::Ipv6(addr)) => {
-            !(addr.is_loopback()
-                || addr.is_unspecified()
-                || addr.is_unique_local()
-                || addr.is_unicast_link_local()
-                || addr.is_multicast())
-        }
-        Some(Host::Domain(host)) => {
-            let host = host.trim_end_matches('.').to_ascii_lowercase();
-            matches!(
-                host.as_str(),
-                "updates.push.services.mozilla.com"
-                    | "fcm.googleapis.com"
-                    | "android.googleapis.com"
-                    | "web.push.apple.com"
-            ) || host.ends_with(".notify.windows.com")
-        }
-        None => false,
-    }
 }
 
 async fn permission_allowed(
@@ -5338,102 +3496,6 @@ async fn apply_muc_owner_config(
     }
 
     Ok(())
-}
-
-async fn handle_command_iq(
-    request_iq: &xmpp_parsers::iq::Iq,
-    state: &WebSocketState,
-    domain: &str,
-    extensions_domain: &str,
-    authenticated_session: &Option<Session>,
-    bound_jid: Option<&FullJid>,
-) -> Vec<String> {
-    let sender_jid: Jid = match bound_jid.cloned().map(Jid::from) {
-        Some(jid) => jid,
-        None => {
-            return vec![build_xmpp_error_response(
-                request_iq,
-                XmppError::not_authorized(Some("Authenticated session required".to_string())),
-            )];
-        }
-    };
-
-    let command = match parse_command_from_iq(request_iq) {
-        Ok(command) => command,
-        Err(err) => {
-            return vec![build_xmpp_error_response(
-                request_iq,
-                XmppError::bad_request(Some(format!("Invalid command request: {err}"))),
-            )];
-        }
-    };
-
-    let node = command.node.clone();
-    let target = request_iq
-        .to
-        .as_ref()
-        .map(|jid| jid.to_bare().to_string())
-        .unwrap_or_else(|| domain.to_string());
-    let extension_command = is_extension_command_node(&node);
-    let allowed_target = if extension_command {
-        Some(extensions_domain)
-    } else {
-        Some(domain)
-    };
-    if Some(target.as_str()) != allowed_target {
-        return vec![build_xmpp_error_response(
-            request_iq,
-            XmppError::service_unavailable(Some(
-                "Command is not available on this service".to_string(),
-            )),
-        )];
-    }
-    let session_id = command.session_id.clone();
-    let ctx = CommandContext {
-        from: sender_jid,
-        authenticated_user_id: authenticated_session
-            .as_ref()
-            .map(|session| session.user_id.clone()),
-        iq: request_iq.clone(),
-        command,
-    };
-
-    let result = state.deps.protocol.command_registry.dispatch(ctx).await;
-    let response_command = match result {
-        CommandResult::Executing {
-            form,
-            session_id,
-            notes,
-        } => {
-            let mut command = Command::new(node.clone());
-            command.status = Some(CommandStatus::Executing);
-            command.session_id = Some(session_id);
-            command.form = Some(form);
-            command.notes = notes;
-            command
-        }
-        CommandResult::Completed { form, notes } => {
-            let mut command = Command::new(node.clone());
-            command.status = Some(CommandStatus::Completed);
-            command.session_id = session_id;
-            command.form = form;
-            command.notes = notes;
-            command
-        }
-        CommandResult::Canceled { notes } => {
-            let mut command = Command::new(node.clone());
-            command.status = Some(CommandStatus::Canceled);
-            command.session_id = session_id;
-            command.notes = notes;
-            command
-        }
-        CommandResult::Error(err) => return vec![build_xmpp_error_response(request_iq, err)],
-    };
-
-    vec![iq_to_xml(build_command_result(
-        request_iq,
-        &response_command,
-    ))]
 }
 
 #[cfg(test)]
